requirements.txt
langchain
langchain_core
langchain_groq
pandas
PyPDF2
python-dotenv
streamlit
streamlit

